server/pyproject.toml
[tool.poetry]
name = "langserve_rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-openai = "^0.1.7"
langchain-postgres = "^0.0.6"
unstructured = {extras = ["pdf"], version = "^0.14.0"}
langchain-community = "^0.2.0"
langchain = "^0.2.0"
langchain-experimental = "^0.0.59"
langchainhub = "^0.1.15"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"
sqlalchemy = "^2.0.30"
fastapi-nextauth-jwt = "^2.0.0"
firebase-admin = "^6.5.0"
ruff = "^0.4.8"
python-poppler = "^0.4.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
